fix always running download ollama

+++ docker/hri/scripts/download-model.sh
@@ -190,33 +190,35 @@ if ask_for_model ei-kws 7; then
     download_ei_model "kws" "${EI_API_KEY_KWS:-}" "1338"
 fi
 
-# Detect available image
-if docker images | grep -q "dustynv/ollama"; then
-    IMAGE="dustynv/ollama:0.6.8-r36.4"
-    COMMAND="ollama serve"
-elif docker images | grep -q "ollama/ollama"; then
-    IMAGE="ollama/ollama"
-    COMMAND=""
-else
-    echo "Error: No compatible Ollama image found. Pulling the default image..."
-    docker pull ollama/ollama:latest
-    IMAGE="ollama/ollama"
-    COMMAND=""
+if ask_for_model qwen3 3 || ask_for_model nomic-embed-text 4; then
+    # Detect available image
+    if docker images | grep -q "dustynv/ollama"; then
+        IMAGE="dustynv/ollama:0.6.8-r36.4"
+        COMMAND="ollama serve"
+    elif docker images | grep -q "ollama/ollama"; then
+        IMAGE="ollama/ollama"
+        COMMAND=""
+    else
+        echo "Error: No compatible Ollama image found. Pulling the default image..."
+        docker pull ollama/ollama:latest
+        IMAGE="ollama/ollama"
+        COMMAND=""
+    fi
+
+    echo "Running: docker run -d --rm --runtime=nvidia -v \"$SCRIPT_DIR\":/ollama -e OLLAMA_MODELS=/ollama $IMAGE $COMMAND"
+
+    # Don't quote $COMMAND to allow for multiple word commands
+    CONTAINER_ID=$(docker run -d --rm --runtime=nvidia -v "$SCRIPT_DIR":/ollama -e OLLAMA_MODELS=/ollama "$IMAGE" $COMMAND)
+
+    if ask_for_model qwen3 3; then
+        docker exec "$CONTAINER_ID" ollama pull qwen3
+    fi
+
+    if ask_for_model nomic-embed-text 4; then
+        docker exec "$CONTAINER_ID" ollama pull nomic-embed-text
+    fi
+
+    docker stop "$CONTAINER_ID"
 fi
-
-echo "Running: docker run -d --rm --runtime=nvidia -v \"$SCRIPT_DIR\":/ollama -e OLLAMA_MODELS=/ollama $IMAGE $COMMAND"
-
-# Don't quote $COMMAND to allow for multiple word commands
-CONTAINER_ID=$(docker run -d --rm --runtime=nvidia -v "$SCRIPT_DIR":/ollama -e OLLAMA_MODELS=/ollama "$IMAGE" $COMMAND)
-
-if ask_for_model qwen3 3; then
-    docker exec "$CONTAINER_ID" ollama pull qwen3
-fi
-
-if ask_for_model nomic-embed-text 4; then
-    docker exec "$CONTAINER_ID" ollama pull nomic-embed-text
-fi
-
-docker stop "$CONTAINER_ID"
 
 echo "All selected models downloaded."
